+++ pysal/contrib/pdio/__init__.py
@@ -1,0 +1,4 @@
+import dbf as dbio
+import shp as shio
+import file as fio
+from file import read_files, write_files
