+++ pysal/contrib/pdio/dbf.py
