+++ pysal/contrib/pdio/shp.py
