--- pysal/contrib/pdutilities/__init__.py
@@ -1,4 +1,0 @@
-import dbf_utilities as dbio
-import shp_utilities as shio
-import file_utilities as fio
-from file_utilities import read_files, write_files
